# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: CultureMatch API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /auth/magic/start:
    post:
      summary: Start email magic link login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicStartRequest"
      responses:
        "200":
          description: Magic link request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicStartResponse"
  /auth/verify:
    post:
      summary: Verify a magic link token and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicVerifyRequest"
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/magic/verify:
    post:
      summary: Verify a magic link token and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MagicVerifyRequest"
                - type: object
                  properties:
                    email:
                      type: string
                      format: email
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/google:
    post:
      summary: Verify a Google ID token and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthRequest"
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/apple:
    post:
      summary: Verify an Apple identity token and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleAuthRequest"
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/refresh:
    post:
      summary: Rotate refresh token and issue a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/logout:
    post:
      summary: Revoke active refresh tokens for the session user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: Logout completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/oauth/{provider}:
    post:
      summary: Legacy OAuth provider endpoint
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [apple, google]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerId]
              properties:
                providerId:
                  type: string
                email:
                  type: string
                  format: email
                deviceId:
                  type: string
                  default: mobile
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /chat/conversations:
    get:
      summary: Chat list with unread counts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationListResponse"
  /chat/conversations/{id}/messages:
    get:
      summary: Conversation messages
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
    post:
      summary: Send a conversation message
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreateRequest"
      responses:
        "201":
          description: Message
  /chat/icebreakers:
    get:
      summary: Cultural icebreaker prompts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: conversationId
          schema:
            type: string
            format: uuid
        - in: query
          name: language
          schema:
            type: string
            default: en
      responses:
        "200":
          description: Icebreakers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
...
  /media/presign:
    post:
      summary: Create a presigned upload URL for a profile photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaPresignRequest"
      responses:
        "200":
          description: Presigned upload target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPresignResponse"
  /media/photos:
    post:
      summary: Register an uploaded profile photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoCreateRequest"
      responses:
        "201":
          description: Photo
  /media/photos/{id}:
    delete:
      summary: Delete a profile photo
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Photo deleted
        "404":
          description: Photo not found
  /media/photos/reorder:
    patch:
      summary: Reorder profile photos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [id, order_index]
                properties:
                  id:
                    type: string
                    format: uuid
                  order_index:
                    type: integer
                    minimum: 0
      responses:
        "200":
          description: Order saved
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
    CursorParam:
      in: query
      name: cursor
      required: false
      schema:
        type: string
      description: Pagination cursor (ISO timestamp from last item's created_at)
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        nextCursor:
          type: string
          nullable: true
    MagicStartRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    MagicStartResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string
        devMagicLink:
          type: string
          description: Present outside production only
    MagicVerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 20
        deviceId:
          type: string
          default: mobile
    GoogleAuthRequest:
      type: object
      anyOf:
        - required: [token]
        - required: [idToken]
      properties:
        token:
          type: string
        idToken:
          type: string
        deviceId:
          type: string
          default: mobile
    AppleAuthRequest:
      type: object
      anyOf:
        - required: [token]
        - required: [identityToken]
      properties:
        token:
          type: string
        identityToken:
          type: string
        deviceId:
          type: string
          default: mobile
    RefreshRequest:
      type: object
      anyOf:
        - required: [refreshToken]
        - required: [refresh_token]
      properties:
        refreshToken:
          type: string
        refresh_token:
          type: string
        deviceId:
          type: string
          default: mobile
    LogoutRequest:
      type: object
      anyOf:
        - required: [refreshToken]
        - required: [refresh_token]
      properties:
        refreshToken:
          type: string
        refresh_token:
          type: string
    AuthSession:
      type: object
      required: [accessToken, refreshToken, userId]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userId:
          type: string
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        peer_id:
          type: string
          format: uuid
        peer_name:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        last_message_content:
          type: string
          nullable: true
        last_message_type:
          type: string
          enum: [text, image, reaction, icebreaker]
          nullable: true
        last_message_sender_id:
          type: string
          format: uuid
          nullable: true
        last_message_created_at:
          type: string
          format: date-time
          nullable: true
        unread_count:
          type: integer
    ConversationListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
    MessageCreateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        type:
          type: string
          enum: [text, image, reaction, icebreaker]
          default: text
...
    MediaPresignRequest:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        sizeBytes:
          type: integer
          maximum: 10485760
    MediaPresignResponse:
      type: object
      required: [uploadUrl, key, photoUrl, publicUrl]
      properties:
        uploadUrl:
          type: string
        key:
          type: string
        photoUrl:
          type: string
        publicUrl:
          type: string
    PhotoCreateRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        key:
          type: string
        thumbnail_url:
          type: string
          format: uri
        order_index:
          type: integer
          default: 0
        is_primary:
          type: boolean
          default: false
...